refactor(ui): simplify Modal narrowing and extract ModalContext
- Move ModalContext into its own file under core/overlay so it can be
  imported without pulling in the Modal component, and re-export from
  the overlay barrel.
- Collapse the Modal function's narrowing down to a direct discriminant
  check (`props.variant === "slideover" && props.isResizable`), drop
  the `in` guard and `as ResizableSlideoverProps` cast.
- Drop `size` from the resizable slideover path — width is driven by
  drag state, so `data-size` has no effect and passing it through was
  misleading.
