+++ app/src/components/core/dialog/Dialog.tsx
@@ -13,7 +13,7 @@ import { Heading } from "@phoenix/components/core/content";
 import { Icon, Icons } from "@phoenix/components/core/icon";
 import type { FlexProps } from "@phoenix/components/core/layout";
 import { Flex } from "@phoenix/components/core/layout";
-import { ModalContext } from "@phoenix/components/core/overlay/Modal";
+import { ModalContext } from "@phoenix/components/core/overlay/ModalContext";
 import { classNames } from "@phoenix/utils/classNames";
 
 export type DialogProps = AriaDialogProps;
