+++ app/src/components/core/overlay/index.tsx
@@ -1,3 +1,4 @@
 export * from "./PopoverArrow";
 export * from "./Popover";
 export * from "./Modal";
+export * from "./ModalContext";
